refactor(common)!: vacate org.apache.parquet.schema namespace from hudi-common
Moves the three hudi classes squatting the parquet namespace to their
proper homes:

- LogicalTypeParquetAdapter / OriginalTypeParquetAdapter ->
  org.apache.hudi.metadata.stats, next to the ParquetAdapter interface
  they implement. The namespace residency was unnecessary: both use
  only public parquet APIs, and ParquetAdapter.getAdapter() references
  them via class literals (the reflective load exists to defer linkage
  for the parquet <1.11 fallback and is unaffected by the package).
- HoodieSchemaRepair -> org.apache.hudi.common.schema: it imports zero
  parquet types - it is pure HoodieSchema manipulation and belongs with
  the schema utilities it already uses; its test moves alongside.

hudi-common no longer contributes any classes to org.apache.parquet.*.
The hadoop-common parquet-namespace classes (SchemaRepair,
HoodieAvroParquetSchemaConverter) are intentionally untouched - that
module is the parquet-facing layer; its equivalence test now imports
the relocated HoodieSchemaRepair explicitly.

BREAKING CHANGE: source/binary break for external code importing
HoodieSchemaRepair (shipped in 1.2.0) or the adapters; loud failure,
import fix. None are config values or reflection targets by name.
