+++ hudi-common/src/main/java/org/apache/hudi/metadata/stats/LogicalTypeParquetAdapter.java
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-package org.apache.parquet.schema;
+package org.apache.hudi.metadata.stats;
 
-import org.apache.hudi.metadata.stats.ParquetAdapter;
-import org.apache.hudi.metadata.stats.ValueType;
+import org.apache.parquet.schema.LogicalTypeAnnotation;
+import org.apache.parquet.schema.PrimitiveType;
 
 /**
  * Uses LogicalTypeAnnotation to extract value type, precision, and scale
